refactor: `importItemMatcher` util
