+++ lib/loader.js
@@ -5,7 +5,11 @@
 const loaderUtils = require('loader-utils');
 
 const processCss = require('./processCss');
-const { getImportPrefix, compileExports } = require('./utils');
+const {
+  getImportPrefix,
+  placeholderImportItemReplacer,
+  compileExports,
+} = require('./utils');
 
 module.exports = function loader(content, map) {
   const callback = this.async();
@@ -78,20 +82,9 @@ module.exports = function loader(content, map) {
         }, this)
         .join('\n');
 
-      function importItemMatcher(item) {
-        const match = result.importItemRegExp.exec(item);
-        const idx = +match[1];
-        const importItem = result.importItems[idx];
-        const importUrl = importUrlPrefix + importItem.url;
-        return `" + require(${loaderUtils.stringifyRequest(
-          this,
-          importUrl
-        )}).locals[${JSON.stringify(importItem.export)}] + "`;
-      }
-
       cssAsString = cssAsString.replace(
         result.importItemRegExpG,
-        importItemMatcher.bind(this)
+        placeholderImportItemReplacer(this, result, importUrlPrefix)
       );
 
       // helper for ensuring valid CSS strings from requires
@@ -136,7 +129,7 @@ module.exports = function loader(content, map) {
 
       let exportJs = compileExports(
         result,
-        importItemMatcher.bind(this),
+        placeholderImportItemReplacer(this, result, importUrlPrefix),
         options.camelCase
       );
       if (exportJs) {
@@ -170,7 +163,7 @@ module.exports = function loader(content, map) {
         moduleJs = `exports.push([module.id, ${cssAsString}, ""]);`;
       }
 
-      // embed runtime
+      // Embed runtime
       return callback(
         null,
         `${urlEscapeHelper}exports = module.exports = require(${loaderUtils.stringifyRequest(
