+++ lib/localsLoader.js
@@ -5,7 +5,11 @@
 const loaderUtils = require('loader-utils');
 
 const processCss = require('./processCss');
-const { getImportPrefix, compileExports } = require('./utils');
+const {
+  getImportPrefix,
+  placeholderImportItemReplacer,
+  compileExports,
+} = require('./utils');
 
 module.exports = function loader(content) {
   const callback = this.async();
@@ -26,20 +30,9 @@ module.exports = function loader(content) {
       // for importing CSS
       const importUrlPrefix = getImportPrefix(this, options);
 
-      function importItemMatcher(item) {
-        const match = result.importItemRegExp.exec(item);
-        const idx = +match[1];
-        const importItem = result.importItems[idx];
-        const importUrl = importUrlPrefix + importItem.url;
-        return (
-          `" + require(${loaderUtils.stringifyRequest(this, importUrl)})` +
-          `[${JSON.stringify(importItem.export)}] + "`
-        );
-      }
-
       let exportJs = compileExports(
         result,
-        importItemMatcher.bind(this),
+        placeholderImportItemReplacer(this, result, importUrlPrefix, true),
         options.camelCase
       );
       if (exportJs) {
