+++ lib/utils.js
@@ -110,9 +110,36 @@ function getLocalIdent(loaderContext, localIdentName, localName, options) {
     .replace(/^((-?[0-9])|--)/, '_$1');
 }
 
+function placeholderImportItemReplacer(
+  loaderContext,
+  result,
+  importUrlPrefix,
+  onlyLocals = false
+) {
+  return (item) => {
+    const match = result.importItemRegExp.exec(item);
+    const idx = +match[1];
+    const importItem = result.importItems[idx];
+    const importUrl = importUrlPrefix + importItem.url;
+
+    if (onlyLocals) {
+      return `" + require(${loaderUtils.stringifyRequest(
+        loaderContext,
+        importUrl
+      )})[${JSON.stringify(importItem.export)}] + "`;
+    }
+
+    return `" + require(${loaderUtils.stringifyRequest(
+      loaderContext,
+      importUrl
+    )}).locals[${JSON.stringify(importItem.export)}] + "`;
+  };
+}
+
 module.exports = {
   dashesCamelCase,
   compileExports,
   getImportPrefix,
   getLocalIdent,
+  placeholderImportItemReplacer,
 };
